Dev/#464 last state in indexed db

+++ santa-fe/src/app/modules/core/services/GlobalWorkflowIOService.ts
@@ -3,7 +3,7 @@ import { DetachedRouteHandle } from '@angular/router';
 import { Observable, Subscription } from 'rxjs';
 import { Store } from '@ngrx/store';
 
-import { DTOs } from '../models/frontend';
+import { DTOs, AdhocPacks } from '../models/frontend';
 import { UtilityService } from 'Core/services/UtilityService';
 import { DTOService } from 'Core/services/DTOService';
 import {
@@ -14,7 +14,8 @@ import { CoreGlobalWorkflowIndexedDBReady } from 'Core/actions/core.actions';
 import {
   INDEXEDDB_VERSION,
   INDEXEDDB_WORKFLOW_DATABASE_NAME,
-  INDEXEDDB_WORKFLOW_STORE_NAME,
+  INDEXEDDB_WORKFLOW_TABLE_NAME,
+  INDEXEDDB_LAST_STATE_TABLE_NAME,
   ROUTE_REUSE_HANDLER_STORE_SIZE_CAP
 } from 'Core/constants/globalWorkflowConstants.constants';
 
@@ -23,12 +24,11 @@ import {
 export class GlobalWorkflowIOService {
   // given that storing workflow state is the only application of indexedDB in Santa at the moment, there is no need to over-engineer the indexedDB layer in santa, just put it in IOService for now
   private workflowIndexedDBAPI: IDBDatabase;
-  private workflowStore: IDBObjectStore;
-  private workflowIO: IDBTransaction;
   constants = {
     idbVersion: INDEXEDDB_VERSION,
     idbWorkflowDbName: INDEXEDDB_WORKFLOW_DATABASE_NAME,
-    idbWorkflowStoreName: INDEXEDDB_WORKFLOW_STORE_NAME
+    idbWorkflowAllStateTableName: INDEXEDDB_WORKFLOW_TABLE_NAME,
+    idbWorkflowLastStateTableName: INDEXEDDB_LAST_STATE_TABLE_NAME
   }
 
   private currentState: string = 'initialState';
@@ -65,21 +65,21 @@ export class GlobalWorkflowIOService {
       const writableCopy = this.utilityService.deepCopy(targetState);
       writableCopy.data.stateInfo = JSON.stringify(writableCopy.data.stateInfo);
       writableCopy.api = null;
-      this.workflowIO = this.workflowIndexedDBAPI.transaction([this.constants.idbWorkflowStoreName], "readwrite");
-      this.workflowStore = this.workflowIO.objectStore(this.constants.idbWorkflowStoreName);
-      this.workflowIO.onerror = (event) => {
+      const IOTransaction = this.workflowIndexedDBAPI.transaction([this.constants.idbWorkflowAllStateTableName], "readwrite");
+      const IOService = IOTransaction.objectStore(this.constants.idbWorkflowAllStateTableName);
+      IOTransaction.onerror = (event) => {
         console.error('Global Workflow, store state error', event);
       }
-      this.workflowStore.put(writableCopy);
+      IOService.put(writableCopy);
     }
 
     public fetchState(targetUUID: string): Observable<DTOs.GlobalWorkflowStateDTO> {
       return new Observable(subscriber => {
         if (!!targetUUID) {
-          this.workflowIO = this.workflowIndexedDBAPI.transaction([this.constants.idbWorkflowStoreName], "readwrite");
-          this.workflowStore = this.workflowIO.objectStore(this.constants.idbWorkflowStoreName);
-          const request = this.workflowStore.get(targetUUID);
-          this.workflowIO.oncomplete = ((event) => {
+          const IOTransaction = this.workflowIndexedDBAPI.transaction([this.constants.idbWorkflowAllStateTableName], "readwrite");
+          const IOService = IOTransaction.objectStore(this.constants.idbWorkflowAllStateTableName);
+          const request = IOService.get(targetUUID);
+          IOTransaction.oncomplete = ((event) => {
             if (!!request.result && !!request.result.data) {
               const workflowDTO: DTOs.GlobalWorkflowStateDTO = {
                 uuid: request.result.uuid,
@@ -100,7 +100,7 @@ export class GlobalWorkflowIOService {
               subscriber.next(null);
             }
           });
-          this.workflowIO.onerror = ((event) => {
+          IOTransaction.onerror = ((event) => {
             console.error('Global Workflow, retrieve state failure', event, targetUUID);
           });
         } else {
@@ -125,21 +125,79 @@ export class GlobalWorkflowIOService {
         console.log('IDB open request upgrade needed.', newVersionDetectedEvent);
         // reconstruct the database upon version change
         this.workflowIndexedDBAPI = openRequest.result;
-        switch(newVersionDetectedEvent.oldVersion) {
-          case 0:
-            // version 0 means that the client had no database
-            // perform initialization
-            this.upgradeIndexedDB(newVersionDetectedEvent);
-            break;
-          default:
-            window.indexedDB.deleteDatabase(this.constants.idbWorkflowDbName);
-            break;
+
+        if (newVersionDetectedEvent.oldVersion === 0) {
+          // version 0 means that the client had no database
+          // perform initialization
+          this.initializeWorkflowTable();
+          this.initializeLastStateTable();
+        } else {
+          // for all other versions, simply destory the database and reload, which will trigger the "version === 0" condition that rebuilds the database
+          // this is not the most efficient way to handle upgrade but it is error-proof
+          const deleteRequest = window.indexedDB.deleteDatabase(this.constants.idbWorkflowDbName);
+          deleteRequest.onsuccess = (event) => {
+            window.location.reload(true);
+          }
         }
+        this.workflowIndexedDBAPI.onversionchange = (event) => {
+          // versionchange event will trigger on other instances of the application that did not trigger the delete of the database
+          // this is to handle scenarios where users have multiple tabs of the Santa open. What will happen is, one of the tabs triggers the delete, the other tabs will refresh themselves as the delete is completed.
+          this.workflowIndexedDBAPI.close();
+          setTimeout(()=>{
+            // delayed to reload in order to make sure when reload database is already deleted
+            window.location.reload(true);
+          }, 1000);
+        };
+
       }
     }
 
-    private upgradeIndexedDB(newVersionDetectedEvent: IDBVersionChangeEvent) {
-      this.workflowStore = this.workflowIndexedDBAPI.createObjectStore(this.constants.idbWorkflowStoreName, { keyPath: "uuid" });  // this key field has to be the "id" field 
+    private initializeWorkflowTable() {
+      this.workflowIndexedDBAPI.createObjectStore(this.constants.idbWorkflowAllStateTableName, { keyPath: "uuid" });  // this key field has to be the "id" field 
+    }
+
+    private initializeLastStateTable() {
+      this.workflowIndexedDBAPI.createObjectStore(this.constants.idbWorkflowLastStateTableName, { keyPath: "module" });
+    }
+
+    private storeLastState(targetModule: NavigationModule, targetUUID: string) {
+      if (!!this.workflowIndexedDBAPI) {
+        // this if condition serves both as a null-check and a guard for not recording the initial state on app load, because it is unnecessary to store it
+        const IOTransaction = this.workflowIndexedDBAPI.transaction([this.constants.idbWorkflowLastStateTableName], "readwrite");
+        const IOService = IOTransaction.objectStore(this.constants.idbWorkflowLastStateTableName);
+        IOTransaction.onerror = (event) => {
+          console.error('Global Workflow, store last state error', event);
+        }
+        const newEntry: AdhocPacks.GlobalWorkflowLastState = {
+          module: targetModule,
+          stateUUID: targetUUID
+        };
+        IOService.put(newEntry);
+      }
+    }
+
+    public loadLastStates(): Observable<Array<AdhocPacks.GlobalWorkflowLastState>> {
+      return new Observable(subscriber => {
+        const results = [];
+        let expectedNumOfResults = 0;
+        for (let eachModule in NavigationModule) {
+          expectedNumOfResults++;
+        }
+        for (let eachModule in NavigationModule) {
+          const IOTransaction = this.workflowIndexedDBAPI.transaction([this.constants.idbWorkflowLastStateTableName], "readwrite");
+          const IOService = IOTransaction.objectStore(this.constants.idbWorkflowLastStateTableName);
+          const request = IOService.get(eachModule);
+          IOTransaction.oncomplete = ((event) => {
+            results.push(request.result);
+            if (results.length >= expectedNumOfResults) {
+              subscriber.next(results);
+            }
+          });
+          IOTransaction.onerror = ((event) => {
+            console.error('Global Workflow, retrieve state failure', event, eachModule);
+          });
+        }
+      });
     }
 
   // Global Workflow States End
@@ -149,6 +207,7 @@ export class GlobalWorkflowIOService {
     public updateCurrentState(newModule: NavigationModule ,newStateId: string) {
       this.currentModule = newModule;
       this.currentState = newStateId;
+      this.storeLastState(newModule, newStateId)
     }
 
     public attachRouteHandlerToState(targetUUID: string, targetHandler: DetachedRouteHandle) {
